Move _handle_stream up
